+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/link/LinkMySQLToModuleDialog.java
@@ -11,19 +11,27 @@ import com.intellij.openapi.progress.ProgressManager;
 import com.intellij.openapi.project.Project;
 import com.microsoft.azure.toolkit.intellij.common.AzureDialog;
 import com.microsoft.azure.toolkit.intellij.link.mysql.BasicLinkMySQLPanel;
+import com.microsoft.azure.toolkit.intellij.link.mysql.JdbcUrl;
 import com.microsoft.azure.toolkit.intellij.link.mysql.MySQLResourceConfig;
+import com.microsoft.azure.toolkit.intellij.link.po.MySQLResourcePO;
 import com.microsoft.azure.toolkit.lib.common.form.AzureForm;
 import com.microsoft.azure.toolkit.lib.common.task.AzureTask;
 import com.microsoft.azure.toolkit.lib.common.task.AzureTaskManager;
 import com.microsoft.azure.toolkit.lib.link.AzureLinkService;
+import com.microsoft.intellij.AzureMySQLStorage;
 import com.microsoft.tooling.msservices.components.DefaultLoader;
 import com.microsoft.tooling.msservices.serviceexplorer.azure.mysql.MySQLNode;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.Nullable;
 
 import javax.swing.*;
+import java.util.Objects;
 
 public class LinkMySQLToModuleDialog extends AzureDialog<LinkConfig<MySQLResourceConfig, ModuleResourceConfig>> {
+
+    private static final String PROMPT_TITLE = "Azure Explorer";
+    private static final String[] PROMPT_OPTIONS = new String[] {"Yes", "No"};
+    private static final String PROMPT_MESSAGE = "This resource already existed in your local environment. Do you want to override it?";
     private JPanel rootPanel;
     private BasicLinkMySQLPanel basicPanel;
 
@@ -71,12 +79,27 @@ public class LinkMySQLToModuleDialog extends AzureDialog<LinkConfig<MySQLResourc
 
     private void doLink(LinkConfig<MySQLResourceConfig, ModuleResourceConfig> linkConfig, Project project, LinkMySQLToModuleDialog dialog) {
         dialog.close(0);
+        // check to prompt override existing resource or not
+        MySQLResourceConfig resourceConfig = linkConfig.getResource();
+        JdbcUrl jdbcUrl = JdbcUrl.from(resourceConfig.getUrl());
+        String businessUniqueKey = MySQLResourcePO.getBusinessUniqueKey(resourceConfig.getServer().id(), jdbcUrl.getDatabase());
+        MySQLResourcePO existedResourcePO = AzureMySQLStorage.getStorage().getResourceByBusinessUniqueKey(businessUniqueKey);
+        boolean storageResource = true;
+        if (Objects.nonNull(existedResourcePO)) {
+            if (!StringUtils.equals(resourceConfig.getUrl(), existedResourcePO.getUrl()) ||
+                    !StringUtils.equals(resourceConfig.getUsername(), existedResourcePO.getUsername()) ||
+                    resourceConfig.getPasswordConfig().getPasswordSaveType() != existedResourcePO.getPasswordSave()) {
+                storageResource = DefaultLoader.getUIHelper().showConfirmation(PROMPT_MESSAGE, PROMPT_TITLE, PROMPT_OPTIONS, null);
+            }
+        }
+        // link in background
+        boolean finalStorageResource = storageResource;
         final Runnable runnable = () -> {
             final ProgressIndicator indicator = ProgressManager.getInstance().getProgressIndicator();
             indicator.setIndeterminate(true);
-            AzureLinkService.getInstance().link(project, linkConfig);
+            AzureLinkService.getInstance().link(project, linkConfig, finalStorageResource);
         };
-        String progressMessage = "Linking Azure Database for MySQL with Module...";
+        String progressMessage = "Connecting Azure Database for MySQL with Module...";
         final AzureTask task = new AzureTask(null, progressMessage, false, runnable);
         AzureTaskManager.getInstance().runInBackground(task);
     }